Renamed some js files to jsx and modified index.html accordingly to fix vite build errors.
