+++ frontend/index.html
@@ -2,7 +2,6 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <!-- Vite handles favicon paths from the root directory automatically -->
     <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
@@ -11,24 +10,13 @@
       content="Web site for FileShare Plus"
     />
     <link rel="apple-touch-icon" href="/logo192.png" />
-    <!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
     <link rel="manifest" href="/manifest.json" />
-    <!--
-      The %PUBLIC_URL% placeholders have been removed as they are not used by Vite.
-      Vite serves static assets from the project root.
-    -->
     <title>FileShare Plus</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <!--
-      This script tag is the entry point for your Vite application.
-      It points to your src/index.js file.
-    -->
-    <script type="module" src="/src/index.js"></script>
+    <!-- CORRECTED: This now points to the new .jsx file -->
+    <script type="module" src="/src/index.jsx"></script>
   </body>
 </html>
